python/requirements.txt
fastapi
uvicorn
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
opentelemetry-instrumentation
opentelemetry-distro
